Fix depending on kt libraries from aar workaround
android_library's implicit aar doesn't flatten its transitive
dependencies. When using the kotlin rules, the kt_android_library rule
creates a few underlying libraries, because of this the classes.jar in
the aar we built was empty. This change separately builds the underlying
kt.jar file, and replaces the aar's classes.jar with the kotlin one.

Signed-off-by: Keith Smiley <keithbsmiley@gmail.com>